Make ThreadPoolExecutor have only 1 worker.
Previously, the ThreadPoolExecutor had the default number of workers,
which was 5 * os.n_cpu(). However, each worker held on to stale
references of the entire file queue, which resulted in massive memory
leaks. These stale references are reclaimed when the thread gets a new
job, so changing the number of workers to 1 fixes the memory leak.